Login

+++ app/src/main/java/es/iescarrillo/android/myfirstappfirebase/activities/InsertOrEditPersonActivity.java
@@ -1,6 +1,6 @@
 package es.iescarrillo.android.myfirstappfirebase.activities;
 
-import android.content.DialogInterface;
+import android.app.AlertDialog;
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
@@ -10,11 +10,18 @@ import android.widget.EditText;
 import android.widget.Toast;
 
 import androidx.activity.EdgeToEdge;
-import androidx.appcompat.app.AlertDialog;
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
+
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.Firebase;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 
 import es.iescarrillo.android.myfirstappfirebase.R;
 import es.iescarrillo.android.myfirstappfirebase.models.Person;
@@ -24,7 +31,7 @@ public class InsertOrEditPersonActivity extends AppCompatActivity {
 
     // Componentes de la pantalla
     private Button btnSave, btnCancel, btnDelete;
-    private EditText etName, etSurname, etAge;
+    private EditText etName, etSurname, etAge, etEmail, etPassword;
 
     // Variables que recibimos en el Intent
     private Person personEdit;
@@ -83,10 +90,8 @@ public class InsertOrEditPersonActivity extends AppCompatActivity {
                     startActivity(intent);
 
                 } else {
-                    // Creamos la nueva persona y seteamos los valores
+                    // Creamos la nueva persona
                     Person person = new Person();
-                    person.setName(etName.getText().toString());
-                    person.setSurname(etSurname.getText().toString());
 
                     if (etName.getText().toString().isBlank()) {
                         Toast.makeText(InsertOrEditPersonActivity.this, "Name is blank", Toast.LENGTH_SHORT).show();
@@ -96,16 +101,46 @@ public class InsertOrEditPersonActivity extends AppCompatActivity {
                         Toast.makeText(InsertOrEditPersonActivity.this, "Surname is blank", Toast.LENGTH_SHORT).show();
                         return;
                     }
+                    if (etEmail.getText().toString().isBlank()) {
+                        Toast.makeText(InsertOrEditPersonActivity.this, "Email is blank", Toast.LENGTH_SHORT).show();
+                        return;
+                    }
+                    if (etPassword.getText().toString().isBlank()) {
+                        Toast.makeText(InsertOrEditPersonActivity.this, "Password is blank", Toast.LENGTH_SHORT).show();
+                        return;
+                    }
 
                     if (!etAge.getText().toString().isBlank())
                         person.setAge(Integer.valueOf(etAge.getText().toString()));
 
-                    String idPerson = personService.insert(person);
-                    Toast.makeText(InsertOrEditPersonActivity.this, "Person with id " + idPerson + " inserted", Toast.LENGTH_SHORT).show();
-                    Log.i("Person id", idPerson);
+                    FirebaseAuth mAuth = FirebaseAuth.getInstance();
+                    mAuth.createUserWithEmailAndPassword(etEmail.getText().toString(), etPassword.getText().toString()).addOnCompleteListener(InsertOrEditPersonActivity.this, new OnCompleteListener<AuthResult>() {
+                        @Override
+                        public void onComplete(@NonNull Task<AuthResult> task) {
+                            if(task.isSuccessful()){
+                                FirebaseUser user = mAuth.getCurrentUser();
 
-                    Intent intent = new Intent(InsertOrEditPersonActivity.this, MainActivity.class);
-                    startActivity(intent);
+                                // Seteamos los valores con los introducidos en la pantalla y los que devuelve la autenticación
+                                person.setName(etName.getText().toString());
+                                person.setSurname(etSurname.getText().toString());
+                                person.setEmail(user.getEmail());
+                                person.setUid(user.getUid());
+                                person.setProvider(user.getProviderId());
+
+                                String idPerson = personService.insert(person);
+                                Toast.makeText(InsertOrEditPersonActivity.this, "Person with id " + idPerson + " inserted", Toast.LENGTH_SHORT).show();
+                                Log.i("Person id", idPerson);
+
+                                Intent intent = new Intent(InsertOrEditPersonActivity.this, MainActivity.class);
+                                startActivity(intent);
+
+                            } else {
+                                Toast.makeText(InsertOrEditPersonActivity.this, "Registration failed.",
+                                        Toast.LENGTH_SHORT).show();
+                            }
+                        }
+                    });
+
                 }
             }
         });
@@ -113,12 +148,22 @@ public class InsertOrEditPersonActivity extends AppCompatActivity {
         btnDelete.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                personService.delete(personEdit);
-                Toast.makeText(getApplicationContext(), "Successfully removed", Toast.LENGTH_SHORT).show();
+                FirebaseAuth mAuth = FirebaseAuth.getInstance();
+                mAuth.getCurrentUser().delete().addOnCompleteListener(InsertOrEditPersonActivity.this, new OnCompleteListener<Void>() {
+                    @Override
+                    public void onComplete(@NonNull Task<Void> task) {
+                        if(task.isSuccessful()){
+                            personService.delete(personEdit);
+                            Toast.makeText(getApplicationContext(), "Successfully removed", Toast.LENGTH_SHORT).show();
 
-                Intent intent = new Intent(InsertOrEditPersonActivity.this, MainActivity.class);
-                startActivity(intent);
-                finish();
+                            Intent intent = new Intent(InsertOrEditPersonActivity.this, MainActivity.class);
+                            startActivity(intent);
+                            finish();
+                        } else{
+                            Toast.makeText(getApplicationContext(), "An error has occurred", Toast.LENGTH_SHORT).show();
+                        }
+                    }
+                });
             }
         });
     }
@@ -131,6 +176,8 @@ public class InsertOrEditPersonActivity extends AppCompatActivity {
         etName = findViewById(R.id.etName);
         etSurname = findViewById(R.id.etSurname);
         etAge = findViewById(R.id.etAge);
+        etEmail = findViewById(R.id.etEmail);
+        etPassword = findViewById(R.id.etPassword);
 
         personService = new PersonService(getApplicationContext());
